requirements.txt
fastapi
uvicorn[standard]
google-genai
faiss-cpu
numpy
python-dotenv
langchain
pydantic
requests
